requirements.txt
altair
attrs
blinker
cachetools
certifi
charset-normalizer
click
extra-streamlit-components
gitdb
GitPython
idna
Jinja2
jsonschema
jsonschema-specifications
markdown-it-py
MarkupSafe
mdurl
narwhals
numpy
packaging
pandas
pillow
protobuf
pyarrow
pydeck
Pygments
python-dateutil
pytz
referencing
requests
rich
rpds-py
six
smmap
streamlit>=1.42.0
streamlit-js-eval
tenacity
toml
tornado
typing_extensions
tzdata
urllib3
python-docx
streamlit-option-menu
python-dotenv
openpyxl
pymongo
Authlib
certifi
beautifulsoup4
fastapi 
uvicorn